+++ CHANGELOG.md
@@ -20,6 +20,10 @@ Introduction of a new module.
 
 - FEATURE: Add `gen_union_spotter()` to allow combining separate spotters into one.
 
+## mini.surround
+
+- FEATURE: Update `add` (`sa`) with ability to replicate left and right parts by respecting `[count]`. In Normal mode two kinds of `[count]` is respected: one for operator (replicates left and right parts) and one for textobject/motion. In Visual mode `[count]` replicates parts.
+
 
 # Version 0.9.0